separate out browser based util to own file
